+++ src/store/modules/users.js
@@ -88,7 +88,7 @@ const users = {
   actions: {
     async ActivateUsers({ dispatch, getters }, { users, _userId }) {
       const updatedUsers = users.map(user => {
-        return { ...user, deactivated: false }
+        return { ...user, is_active: true }
       })
       const nicknames = users.map(user => user.nickname)
       const callApiFn = async() => await activateUsers(nicknames, getters.authHost, getters.token)
@@ -132,7 +132,7 @@ const users = {
     },
     async ApproveUsersAccount({ dispatch, getters }, { users, _userId, _statusId }) {
       const updatedUsers = users.map(user => {
-        return { ...user, approval_pending: false }
+        return { ...user, is_approved: true }
       })
       const nicknames = users.map(user => user.nickname)
       const callApiFn = async() => await approveUserAccount(nicknames, getters.authHost, getters.token)
@@ -149,7 +149,7 @@ const users = {
     },
     async ConfirmUsersEmail({ dispatch, getters }, { users, _userId, _statusId }) {
       const updatedUsers = users.map(user => {
-        return { ...user, confirmation_pending: false }
+        return { ...user, is_confirmed: true }
       })
       const nicknames = users.map(user => user.nickname)
       const callApiFn = async() => await confirmUserEmail(nicknames, getters.authHost, getters.token)
@@ -168,7 +168,7 @@ const users = {
     },
     async DeactivateUsers({ dispatch, getters }, { users, _userId }) {
       const updatedUsers = users.map(user => {
-        return { ...user, deactivated: true }
+        return { ...user, is_active: false }
       })
       const nicknames = users.map(user => user.nickname)
       const callApiFn = async() => await deactivateUsers(nicknames, getters.authHost, getters.token)
@@ -200,7 +200,7 @@ const users = {
         return
       }
       const updatedUsers = users.map(user => {
-        return { ...user, deactivated: true }
+        return { ...user, is_active: false }
       })
       commit('SWAP_USERS', updatedUsers)